@@ -250,7 +250,7 @@ class ArchitectPdfGenerateController extends Controller
             }
 
             if (!empty($tt->SvgImage)) {
-                $svgFile = strpos($tt->SvgImage,'.png') !== false ? URL('/').'/uploads/files/'.$tt->SvgImage : $tt->SvgImage;
+                $svgFile = str_contains((string) $tt->SvgImage,'.png') ? URL('/').'/uploads/files/'.$tt->SvgImage : $tt->SvgImage;
             } else {
                 $svgFile = URL('/').'/uploads/files/door.jpg';
             }
