@@ -20,19 +20,17 @@ use App\Models\SettingIntumescentSeals2;
 
 class InvoiceInExcel implements FromCollection,WithHeadings,WithEvents
 {
-    /**
-    * @return \Illuminate\Support\Collection
-    */
-    protected $id;
-
-    /**
-     * @return \Illuminate\Support\Collection
-     */
-    protected $vid;
-
-    public function __construct($id,$vid) {
-        $this->id = $id;
-        $this->vid = $vid;
+    public function __construct(
+        /**
+         * @return \Illuminate\Support\Collection
+         */
+        protected $id,
+        /**
+         * @return \Illuminate\Support\Collection
+         */
+        protected $vid
+    )
+    {
     }
 
     public function collection()
