@@ -30,10 +30,8 @@ class UserFactory extends Factory
      */
     public function unverified()
     {
-        return $this->state(function (array $attributes): array {
-            return [
-                'email_verified_at' => null,
-            ];
-        });
+        return $this->state(fn(array $attributes): array => [
+            'email_verified_at' => null,
+        ]);
     }
 }
